Revert "Welcome message (#57)"
This reverts commit 27a97872f1ec4e5d32646ad51f0075d57abccc03.
