+++ src/Discord/DiscordEvents.js
@@ -1,7 +1,7 @@
 // get functions used to do things like strip html and replace custom discord emojis with the url to the image
 const { formatMessage } = require("../utils/messageManipulation");
 const CommandHandler = require("./CommandHandler");
-const ReactionRoles = require("./Reaction Roles")
+
 // TODO: move to firebase db
 const ranks = require("../ranks.json");
 
@@ -14,7 +14,6 @@ const eventFiles = fs.readdirSync(eventPath);
 const events = {};
 
 module.exports = (DiscordClient, io, app) => {
-    ReactionRoles(DiscordClient)
     // TODO: move discord events to separate file
     eventFiles.forEach(event => {
         if (event.endsWith(".js")) {
